refactor(webapp): group currency conversion under conversion/, rename tooltip→label
Pipe and helper now describe what they produce (a localized conversion
description string) rather than where they're consumed. The string remains
used as a tooltip but is also valid for aria-label and other display
contexts.
