+++ frontend/projects/webapp/src/app/core/currency/conversion/conversion-label.pipe.ts
@@ -1,6 +1,6 @@
 import { Pipe, type PipeTransform, inject } from '@angular/core';
 import { TranslocoService } from '@jsverse/transloco';
-import { buildConversionTooltip } from './conversion-tooltip';
+import { buildConversionLabel } from './conversion-label';
 
 interface ConversionSource {
   originalAmount?: number | null;
@@ -8,13 +8,13 @@ interface ConversionSource {
   exchangeRate?: number | null;
 }
 
-@Pipe({ name: 'conversionTooltip' })
-export class ConversionTooltipPipe implements PipeTransform {
+@Pipe({ name: 'conversionLabel' })
+export class ConversionLabelPipe implements PipeTransform {
   readonly #transloco = inject(TranslocoService);
 
   transform(source: ConversionSource | null | undefined): string {
     if (!source) return '';
-    return buildConversionTooltip(
+    return buildConversionLabel(
       this.#transloco,
       source.originalAmount,
       source.originalCurrency,
