+++ frontend/projects/webapp/src/app/core/currency/conversion/conversion-label.spec.ts
@@ -4,9 +4,9 @@ import { describe, it, expect, beforeEach } from 'vitest';
 
 import { provideTranslocoForTest } from '@app/testing/transloco-testing';
 
-import { buildConversionTooltip } from './conversion-tooltip';
+import { buildConversionLabel } from './conversion-label';
 
-describe('buildConversionTooltip', () => {
+describe('buildConversionLabel', () => {
   let transloco: TranslocoService;
 
   beforeEach(() => {
@@ -17,33 +17,33 @@ describe('buildConversionTooltip', () => {
   });
 
   it('returns empty string when originalAmount is null', () => {
-    const result = buildConversionTooltip(transloco, null, 'CHF', 1.05);
+    const result = buildConversionLabel(transloco, null, 'CHF', 1.05);
 
     expect(result).toBe('');
   });
 
   it('returns empty string when originalCurrency is null', () => {
-    const result = buildConversionTooltip(transloco, 100, null, 1.05);
+    const result = buildConversionLabel(transloco, 100, null, 1.05);
 
     expect(result).toBe('');
   });
 
   it('with rate uses convertedFromTooltip key and includes the formatted amount and rate', () => {
-    const result = buildConversionTooltip(transloco, 100, 'CHF', 1.05);
+    const result = buildConversionLabel(transloco, 100, 'CHF', 1.05);
 
     expect(result).toContain('100');
     expect(result).toContain('1,05');
   });
 
   it('without rate uses convertedFromTooltipNoRate key and includes the formatted amount', () => {
-    const result = buildConversionTooltip(transloco, 100, 'CHF', null);
+    const result = buildConversionLabel(transloco, 100, 'CHF', null);
 
     expect(result).toContain('100');
     expect(result).not.toContain('taux');
   });
 
   it('formats CHF amount with fr-CH locale', () => {
-    const result = buildConversionTooltip(transloco, 1234.5, 'CHF', null);
+    const result = buildConversionLabel(transloco, 1234.5, 'CHF', null);
 
     expect(result).toContain('CHF');
     // ICU/CLDR fr-CH uses a narrow no-break space for thousands. Assert the digits
@@ -53,7 +53,7 @@ describe('buildConversionTooltip', () => {
   });
 
   it('formats EUR amount with fr-FR locale', () => {
-    const result = buildConversionTooltip(transloco, 1234.5, 'EUR', null);
+    const result = buildConversionLabel(transloco, 1234.5, 'EUR', null);
 
     expect(result).toContain('€');
     // ICU/CLDR fr-FR uses a narrow no-break space for thousands and a comma for
