+++ frontend/projects/webapp/src/app/core/currency/conversion/currency-converter.service.spec.ts
