+++ frontend/projects/webapp/src/app/core/currency/conversion/currency.types.ts
