+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-grid/budget-detail-panel.ts
@@ -16,7 +16,7 @@ import { MatIconModule } from '@angular/material/icon';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
 import { MatTooltipModule } from '@angular/material/tooltip';
 import { type BudgetLine, type Transaction } from 'pulpe-shared';
-import { AppCurrencyPipe, ConversionTooltipPipe } from '@core/currency';
+import { AppCurrencyPipe, ConversionLabelPipe } from '@core/currency';
 import { FeatureFlagsService } from '@core/feature-flags';
 import { UserSettingsStore } from '@core/user-settings';
 import { TranslocoPipe } from '@jsverse/transloco';
@@ -67,7 +67,7 @@ const DETAIL_SEGMENT_COUNT = 12;
     MatSlideToggleModule,
     MatTooltipModule,
     AppCurrencyPipe,
-    ConversionTooltipPipe,
+    ConversionLabelPipe,
     TranslocoPipe,
     DatePipe,
     OriginalAmountLine,
@@ -124,7 +124,7 @@ const DETAIL_SEGMENT_COUNT = 12;
               [displayCurrency]="currency()"
               [tooltipText]="
                 isMultiCurrencyEnabled()
-                  ? (envelope.data | conversionTooltip)
+                  ? (envelope.data | conversionLabel)
                   : ''
               "
             />
@@ -268,7 +268,7 @@ const DETAIL_SEGMENT_COUNT = 12;
                       [originalCurrency]="tx.originalCurrency"
                       [displayCurrency]="currency()"
                       [tooltipText]="
-                        isMultiCurrencyEnabled() ? (tx | conversionTooltip) : ''
+                        isMultiCurrencyEnabled() ? (tx | conversionLabel) : ''
                       "
                     />
                   </div>
