+++ frontend/projects/webapp/src/app/feature/budget/budget-details/budget-grid/budget-grid.ts
@@ -14,7 +14,7 @@ import { MatCardModule } from '@angular/material/card';
 import { MatDialog } from '@angular/material/dialog';
 import { MatIconModule } from '@angular/material/icon';
 import { MatSlideToggleModule } from '@angular/material/slide-toggle';
-import { AppCurrencyPipe, ConversionTooltipPipe } from '@core/currency';
+import { AppCurrencyPipe, ConversionLabelPipe } from '@core/currency';
 import { FeatureFlagsService } from '@core/feature-flags';
 import { FinancialKindDirective } from '@ui/financial-kind';
 import { FinancialKindIndicator } from '@ui/financial-kind-indicator';
@@ -83,7 +83,7 @@ export function groupByKind<T extends { data: { kind: string } }>(
     FinancialKindIndicator,
     FinancialKindDirective,
     OriginalAmountLine,
-    ConversionTooltipPipe,
+    ConversionLabelPipe,
     TransactionActionMenu,
     TransactionLabelPipe,
   ],
@@ -248,9 +248,7 @@ export function groupByKind<T extends { data: { kind: string } }>(
                 [originalCurrency]="item.data.originalCurrency"
                 [displayCurrency]="currency()"
                 [tooltipText]="
-                  isMultiCurrencyEnabled()
-                    ? (item.data | conversionTooltip)
-                    : ''
+                  isMultiCurrencyEnabled() ? (item.data | conversionLabel) : ''
                 "
               />
             </div>
@@ -324,7 +322,7 @@ export function groupByKind<T extends { data: { kind: string } }>(
             [originalCurrency]="item.data.originalCurrency"
             [displayCurrency]="currency()"
             [tooltipText]="
-              isMultiCurrencyEnabled() ? (item.data | conversionTooltip) : ''
+              isMultiCurrencyEnabled() ? (item.data | conversionLabel) : ''
             "
           />
         </div>
